Extract merge catalog pass2 into its own module
Move theirs-only adoption and rootpage remapping out of
doltlite_merge.c into doltlite_merge_pass2.c. Share SchemaRootpageRemap
via doltlite_merge_int.h with the disjoint-schema rebuild path that
stays in the catalog module. Wire main.mk, amalgamation, AGENTS.md, and
require pass2 in the merge size/existence lints.

Verified: lint_layers; merge, index-conflict, schema-merge, conflicts,
and cherry-pick shell suites.
