+++ src/doltlite_merge_pass2.c
@@ -1,0 +1,206 @@
+#ifdef DOLTLITE_PROLLY
+
+#include "doltlite_merge_int.h"
+
+/* Catalog merge pass 2: adopt theirs-only tables/indexes, rootpage remap. */
+
+int mergeCatalogPass2(
+  struct TableEntry *aAnc, int nAnc,
+  struct TableEntry *aOurs, int nOurs,
+  struct TableEntry *aTheirs, int nTheirs,
+  SchemaEntry *aAncSchema, int nAncSchema,
+  SchemaEntry *aOursSchema, int nOursSchema,
+  SchemaEntry *aTheirsSchema, int nTheirsSchema,
+  struct TableEntry *aMerged, int *pnMerged,
+  Pgno *piNextMerged,
+  int bDisjointSchemaChanges,
+  MergeConflictTable *aConflictTables,
+  int nConflictTables,
+  SchemaRootpageRemap **ppaRemap,
+  int *pnRemap,
+  char ***pazReindex,
+  int *pnReindex
+){
+  int i, rc = SQLITE_OK;
+
+  for(i=0; i<nTheirs; i++){
+    const char *zName = aTheirs[i].zName;
+    struct TableEntry *oursEntry;
+
+    if( aTheirs[i].iTable<=1 ) continue;
+
+    if( !zName ){
+      SchemaEntry *pTheirSe = findSchemaEntryByRootpage(
+          aTheirsSchema, nTheirsSchema, aTheirs[i].iTable);
+      if( pTheirSe && pTheirSe->zName && pTheirSe->zType
+       && strcmp(pTheirSe->zType, "index")==0 ){
+        if( hasSchemaConflictObject(aConflictTables, nConflictTables,
+                                    pTheirSe->zName) ){
+          continue;
+        }
+        oursEntry = findCatalogEntryBySchemaObject(
+            aOurs, nOurs, aOursSchema, nOursSchema,
+            pTheirSe->zType, pTheirSe->zName, pTheirSe->zTblName);
+        if( oursEntry ) continue;
+
+        {
+          struct TableEntry *ancEntry = findCatalogEntryBySchemaObject(
+              aAnc, nAnc, aAncSchema, nAncSchema,
+              pTheirSe->zType, pTheirSe->zName, pTheirSe->zTblName);
+          if( !ancEntry ){
+            struct TableEntry newEntry = aTheirs[i];
+            int j, conflict = 0;
+            Pgno oldPg = newEntry.iTable;
+            int forceRemap = bDisjointSchemaChanges;
+            for(j=0; j<*pnMerged; j++){
+              if( aMerged[j].iTable==newEntry.iTable ){
+                conflict = 1;
+                break;
+              }
+            }
+            if( conflict || forceRemap ){
+              SchemaRootpageRemap *aNew;
+              int nOld = *pnRemap;
+              newEntry.iTable = (*piNextMerged)++;
+              aNew = sqlite3_realloc(*ppaRemap,
+                                     (nOld+1)*(int)sizeof(SchemaRootpageRemap));
+              if( !aNew ) return SQLITE_NOMEM;
+              *ppaRemap = aNew;
+              aNew[nOld].oldPg = oldPg;
+              aNew[nOld].newPg = newEntry.iTable;
+              *pnRemap = nOld + 1;
+            }
+            if( newEntry.iTable >= *piNextMerged ) *piNextMerged = newEntry.iTable + 1;
+            aMerged[(*pnMerged)++] = newEntry;
+            /* The adopted tree covers only theirs' rows; ours' row changes
+            ** never touched it. Record the index for a rebuild over the
+            ** merged table once the merged catalog is live. */
+            rc = mergeAppendReindexName(pazReindex, pnReindex, pTheirSe->zName);
+            if( rc!=SQLITE_OK ) return rc;
+          }else if( schemaEntryChangedByName(
+                        aAncSchema, nAncSchema,
+                        aTheirsSchema, nTheirsSchema,
+                        pTheirSe->zName) ){
+            struct TableEntry newEntry = aTheirs[i];
+            int j, conflict = 0;
+            Pgno oldPg = newEntry.iTable;
+            for(j=0; j<*pnMerged; j++){
+              if( aMerged[j].iTable==newEntry.iTable ){
+                conflict = 1;
+                break;
+              }
+            }
+            if( conflict ){
+              SchemaRootpageRemap *aNew;
+              int nOld = *pnRemap;
+              newEntry.iTable = (*piNextMerged)++;
+              aNew = sqlite3_realloc(*ppaRemap,
+                  (nOld+1)*(int)sizeof(SchemaRootpageRemap));
+              if( !aNew ) return SQLITE_NOMEM;
+              *ppaRemap = aNew;
+              aNew[nOld].oldPg = oldPg;
+              aNew[nOld].newPg = newEntry.iTable;
+              *pnRemap = nOld + 1;
+            }
+            if( newEntry.iTable >= *piNextMerged ){
+              *piNextMerged = newEntry.iTable + 1;
+            }
+            aMerged[(*pnMerged)++] = newEntry;
+            rc = mergeAppendReindexName(pazReindex, pnReindex, pTheirSe->zName);
+            if( rc!=SQLITE_OK ) return rc;
+          }else{
+            /* An unmodified index does not override our explicit DROP. */
+          }
+        }
+        continue;
+      }
+
+      if( !doltliteFindTableByNumber(aOurs, nOurs, aTheirs[i].iTable) ){
+        struct TableEntry newEntry = aTheirs[i];
+        int j, conflict = 0;
+        Pgno oldPg = newEntry.iTable;
+        int forceRemap = bDisjointSchemaChanges;
+        for(j=0; j<*pnMerged; j++){
+          if( aMerged[j].iTable==newEntry.iTable ){
+            conflict = 1;
+            break;
+          }
+        }
+        if( conflict || forceRemap ){
+          SchemaRootpageRemap *aNew;
+          int nOld = *pnRemap;
+          Pgno newPg = *piNextMerged;
+          aNew = sqlite3_realloc(*ppaRemap,
+                                 (nOld+1)*(int)sizeof(SchemaRootpageRemap));
+          if( !aNew ) return SQLITE_NOMEM;
+          *ppaRemap = aNew;
+          aNew[nOld].oldPg = oldPg;
+          aNew[nOld].newPg = newPg;
+          *pnRemap = nOld + 1;
+          newEntry.iTable = newPg;
+          (*piNextMerged)++;
+        }
+        if( newEntry.iTable >= *piNextMerged ) *piNextMerged = newEntry.iTable + 1;
+        aMerged[(*pnMerged)++] = newEntry;
+      }else if( bDisjointSchemaChanges ){
+        struct TableEntry *oursIdx = doltliteFindTableByNumber(aOurs, nOurs, aTheirs[i].iTable);
+        struct TableEntry *ancIdx = doltliteFindTableByNumber(aAnc, nAnc, aTheirs[i].iTable);
+        if( oursIdx && !ancIdx
+         && (prollyHashCompare(&oursIdx->root, &aTheirs[i].root)!=0
+             || prollyHashCompare(&oursIdx->schemaHash, &aTheirs[i].schemaHash)!=0) ){
+          struct TableEntry newEntry = aTheirs[i];
+          SchemaRootpageRemap *aNew;
+          int nOld = *pnRemap;
+          Pgno newPg = *piNextMerged;
+          aNew = sqlite3_realloc(*ppaRemap,
+                                 (nOld+1)*(int)sizeof(SchemaRootpageRemap));
+          if( !aNew ) return SQLITE_NOMEM;
+          *ppaRemap = aNew;
+          aNew[nOld].oldPg = aTheirs[i].iTable;
+          aNew[nOld].newPg = newPg;
+          *pnRemap = nOld + 1;
+          newEntry.iTable = newPg;
+          (*piNextMerged)++;
+          aMerged[(*pnMerged)++] = newEntry;
+        }
+      }
+      continue;
+    }
+
+    if( hasSchemaConflictObject(aConflictTables, nConflictTables, zName) ){
+      continue;
+    }
+    oursEntry = doltliteFindTableByName(aOurs, nOurs, zName);
+    if( oursEntry ) continue;
+
+    {
+      struct TableEntry *ancEntry = doltliteFindTableByName(aAnc, nAnc, zName);
+      if( !ancEntry ){
+
+        struct TableEntry newEntry = aTheirs[i];
+        int forceRemap = bDisjointSchemaChanges;
+        {
+          int j, conflict = 0;
+          for(j=0; j<*pnMerged; j++){
+            if( aMerged[j].iTable==newEntry.iTable ){ conflict = 1; break; }
+          }
+          if( conflict || forceRemap ) newEntry.iTable = (*piNextMerged)++;
+        }
+        if( newEntry.iTable >= *piNextMerged ) *piNextMerged = newEntry.iTable + 1;
+
+        newEntry.zName = sqlite3_mprintf("%s", zName);
+        aMerged[(*pnMerged)++] = newEntry;
+      }else{
+
+        int theirsChanged = prollyHashCompare(&aTheirs[i].root, &ancEntry->root)!=0;
+        if( theirsChanged ){
+          return SQLITE_ERROR;
+        }
+
+      }
+    }
+  }
+  return SQLITE_OK;
+}
+
+#endif
